Extend OpCodes to short and add location info to character list

+++ common/src/main/java/pro/velovec/inferno/reborn/common/constants/CommonErrorCodes.java
@@ -2,11 +2,11 @@ package pro.velovec.inferno.reborn.common.constants;
 
 public class CommonErrorCodes {
 
-    public static final byte SUCCESS = 0x00;
+    public static final short SUCCESS = 0x00;
 
-    public static final byte AUTH_ERROR = 0x01;
-    public static final byte AUTH_REQUIRED = 0x02;
+    public static final short AUTH_ERROR = 0x01;
+    public static final short AUTH_REQUIRED = 0x02;
 
-    public static final byte UNKNOWN_OPCODE = 0x7D;
-    public static final byte SERVER_ERROR = 0x7F;
+    public static final short UNKNOWN_OPCODE = 0x7D;
+    public static final short SERVER_ERROR = 0x7F;
 }
