+++ realmd/src/main/java/pro/velovec/inferno/reborn/realmd/constants/RealmErrorCodes.java
@@ -2,14 +2,14 @@ package pro.velovec.inferno.reborn.realmd.constants;
 
 public class RealmErrorCodes {
 
-    public static final byte ALREADY_EXISTS = 0x01;
-    public static final byte CHARACTER_EXISTS = 0x01;
+    public static final short ALREADY_EXISTS = 0x01;
+    public static final short CHARACTER_EXISTS = 0x01;
 
-    public static final byte AUTH_INVALID = 0x02;
-    public static final byte CHARACTER_NOT_FOUND = 0x02;
+    public static final short AUTH_INVALID = 0x02;
+    public static final short CHARACTER_NOT_FOUND = 0x02;
 
-    public static final byte USER_BANNED = 0x03;
-    public static final byte CHARACTER_DELETED = 0x03;
+    public static final short USER_BANNED = 0x03;
+    public static final short CHARACTER_DELETED = 0x03;
 
     private RealmErrorCodes() {
         // Prevent class instantiation
