+++ worldd/src/main/java/pro/velovec/inferno/reborn/worldd/map/WorldCell.java
@@ -38,7 +38,7 @@ public class WorldCell {
         return subscribers.contains(object);
     }
 
-    public synchronized void onEvent(WorldObject source, byte eventType, ByteConvertible eventData) {
+    public synchronized void onEvent(WorldObject source, short eventType, ByteConvertible eventData) {
         subscribers.parallelStream()
             .filter(subscriber -> !subscriber.equals(source))
             .forEach(subscriber -> subscriber.onEvent(
